Added new ProxyGlideModule to load images from proxy
Extracted the OkHttpClient initializiation to factory
Changed the AlertData geometry from Point to generic Geometry
