+++ dependencies.gradle
@@ -54,10 +54,11 @@ ext {
             daggerCompiler       : 'com.google.dagger:dagger-compiler:2.2',
             javaxAnnotations     : 'javax.annotation:jsr250-api:1.0',
             javaxInject          : 'javax.inject:javax.inject:1',
-            autoFactoryCompiler  : "com.google.auto.factory:auto-factory:1.0-beta3",
+            autoFactoryCompiler  : 'com.google.auto.factory:auto-factory:1.0-beta5',
 
             // Glide
             glide                : 'com.github.bumptech.glide:glide:3.7.0',
+            glideOkHttp          : 'com.github.bumptech.glide:okhttp3-integration:1.4.0@aar',
 
             //log4j
             log4jLogging         : 'de.mindpipe.android:android-logging-log4j:1.0.3',
